kill unneccessary cruft/old ideas
we're not using an App Shortcut because the app itself suffices
(besides, the one I have doesn't seem to work for some reason)

that settings page was added back when I thought there was no
way to avoid having the app show some sort of screen.

that BroadcastReceiver was... well, I'm not even sure where I was
going with that.  I think it was supposed to be the eventual target
of my Shortcut, but then I learned that those have to point
to Activities.

android-studio's "Safe Delete" makes me shit rainbows of glee.